local_mcp_00/sample-mcp-clients/pyproject.toml
[project]
name = "sample-mcp-clients"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.12.4",
    "mcp>=1.18.0",
    "openai>=2.5.0",
    "python-dotenv>=1.1.1",
]
